+++ Src/Private/Get-AbrVbrBackupRepository.ps1
@@ -166,6 +166,7 @@ function Get-AbrVbrBackupRepository {
                                                 $OutObj += [pscustomobject]$inobj
 
                                                 if ($HealthCheck.Infrastructure.BR) {
+                                                    $OutObj | Where-Object { $_.'Immutability Supported' -eq 'Yes' } | Set-Style -Style OK -Property 'Immutability Supported'
                                                     $OutObj | Where-Object { $_.'Immutability Supported' -eq 'Yes' -and $_.'Immutability Enabled' -eq 'No' } | Set-Style -Style Warning -Property 'Immutability Enabled'
                                                 }
 
@@ -178,6 +179,11 @@ function Get-AbrVbrBackupRepository {
                                                     $TableParams['Caption'] = "- $($TableParams.Name)"
                                                 }
                                                 $OutObj | Table @TableParams
+
+                                                if (($HealthCheck.Infrastructure.BestPractice) -and ($OutObj | Where-Object { $_.'Immutability Supported' -eq 'Yes' -and $_.'Immutability Enabled' -eq 'No' })) {
+                                                    Paragraph "Health Check:" -Italic -Bold -Underline
+                                                    Paragraph "Best Practice: Veeam recommend to implement Immutability where it is supported. It is done for increased security: immutability protects your data from loss as a result of attacks, malware activity or any other injurious actions." -Italic -Bold
+                                                }
                                             }
                                         }
                                         catch {