+++ Src/Private/Get-AbrVbrConfigurationBackupSetting.ps1
@@ -80,6 +80,20 @@ function Get-AbrVbrConfigurationBackupSetting {
                         $TableParams['Caption'] = "- $($TableParams.Name)"
                     }
                     $OutObj | Table @TableParams
+                    if ($HealthCheck.Infrastructure.BestPractice) {
+                        if ($OutObj | Where-Object { $_.'Encryption Enabled' -like 'No' -or $_.'Run Job Automatically' -like 'No' -or  $_.'Enabled' -like 'No' }) {
+                            Paragraph "Health Check:" -Italic -Bold -Underline
+                            if ($OutObj | Where-Object { $_.'Encryption Enabled' -like 'No'} ) {
+                                Paragraph "Best Practice: Whenever possible, enable configuration backup encryption." -Italic -Bold
+                            }
+                            if ($OutObj | Where-Object { $_.'Run Job Automatically' -like 'No'}) {
+                                Paragraph "Best Practice: It`s a recommended best practice to activate the 'Run job automatically' option of the Backup Configuration job." -Italic -Bold
+                            }
+                            if ($OutObj | Where-Object { $_.'Enabled' -like 'No'}) {
+                                Paragraph "Best Practice: It`s a recommended best practice to enable the Backup Configuration j ob" -Italic -Bold
+                            }
+                        }
+                    }
                 }
             }
         }